Enhance scene transport control and settings management
- Introduced a new 'use_transport' feature in the scene structure, allowing animation to follow transport controls or run continuously.
- Updated scene initialization to default 'use_transport' to false.
- Added console commands for listing and setting module settings, enabling dynamic configuration of module behaviors.
- Enhanced JSON schema and scene files to include 'use_transport' property for better scene management.
- Improved tempo module to respect the 'use_transport' setting, ensuring animations behave correctly based on user preferences.

+++ components/event_bus/include/event_bus.h
@@ -200,7 +200,7 @@ typedef struct {
     
     // Tempo event data
     struct {
-      uint8_t bpm;           // Current tempo in BPM (30-250)
+      uint16_t bpm;          // Current tempo in BPM (20-300)
     } tempo;
     
     // Scene event data
